refractor injiweb changes to v1 to have common version for mobile and web Signed-off-by: Challarao <challarao@beehyv.com>
