+++ src/main/java/io/mosip/mimoto/service/IssuersService.java
@@ -2,14 +2,21 @@ package io.mosip.mimoto.service;
 
 import io.mosip.mimoto.dto.IssuerDTO;
 import io.mosip.mimoto.dto.IssuersDTO;
+import io.mosip.mimoto.dto.v2.IssuerSupportedCredentialsResponse;
 import io.mosip.mimoto.exception.ApiNotAccessibleException;
 
+import java.io.ByteArrayInputStream;
 import java.io.IOException;
 
 public interface IssuersService {
-    IssuersDTO getAllIssuers() throws ApiNotAccessibleException, IOException;
+    IssuersDTO getAllIssuers(String search) throws ApiNotAccessibleException, IOException;
 
     IssuersDTO getAllIssuersWithAllFields() throws ApiNotAccessibleException, IOException;
 
     IssuerDTO getIssuerConfig(String issuerId) throws ApiNotAccessibleException, IOException;
+
+    IssuerSupportedCredentialsResponse getCredentialsSupported(String issuerId, String search) throws ApiNotAccessibleException, IOException;
+
+    ByteArrayInputStream generatePdfForVerifiableCredentials(String token, String issuerId, String credentialsSupportedId) throws Exception;
+
 }
