+++ src/main/resources/application-local.properties
@@ -200,8 +200,8 @@ wallet.binding.partner.api.key=Aci9jg28B8mO_LDfDXo3ZTp5_HKgEMun2tYyHCa1e8k
 #wallet.binding.partner.api.key=14378
 
 # OpenID
-mosip.openid.issuers=v2-issuers-config.json
-mosip.openid.v2.issuer.credentialSupported=/wellKnownIssuer/Insurance.json
+mosip.openid.issuers=mimoto-issuers-config.json
+mosip.openid.v2.issuer.credentialSupported=Insurance.json
 mosip.openid.v2.htmlTemplate=CredentialTemplate.html
 mosip.oidc.client.assertion.type=urn:ietf:params:oauth:client-assertion-type:jwt-bearer
 mosip.oidc.p12.filename=oidckeystore.p12
